% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApproxF.R
\name{ApproxF}
\alias{ApproxF}
\title{Calculate ANOVA with approximate F value}
\usage{
ApproxF(SS.table,approx.name)
}
\arguments{
\item{SS.table}{result from EMSanova}

\item{approx.name}{rowname in SS.table to calculate approximate F value for the test.}
}
\description{
Calculate ANOVA with approximate F value
}
\examples{
G<-rep(rep(1:3,each=2),6)
O<-rep(rep(c("A","B","C"),each=6),2)
D<-rep(1:2,each=18)
Y<-c(0.38,0.4,0.63,0.59,0.76,0.78,0.39,0.41,0.72,0.7,
     0.95,0.96,0.45,0.4,0.78,0.79,1.03,1.06,0.4,0.4,
     0.68,0.66,0.86,0.82,0.39,0.43,0.77,0.76,0.86,0.85,
     0.41,0.4,0.85,0.84,1.01,0.98)
tot.data<-data.frame(Y=Y,G=G,O=O,D=D)
anova.result<-EMSanova(Y~G+O+D,data=tot.data,
                        type=c("F","R","R"))
anova.result                         
ApproxF(SS.table=anova.result,approx.name="G") 
EMSanova(Y~G+O+D,data=tot.data,
          type=c("F","R","R"),
          approximate=TRUE)                      
       
}

