% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMSApp.R
\name{EMSaovApp}
\alias{EMSaovApp}
\title{Shiny App for the analysis of variance in various experimental designs}
\usage{
EMSaovApp()
}
\description{
Shiny App for the analysis of variance in various experimental designs
}
\examples{
#EMSaovApp()
}

