% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PooledANOVA.R
\name{PooledANOVA}
\alias{PooledANOVA}
\title{Pooling nonsignificant interactions to Residuals}
\usage{
PooledANOVA(SS.table,del.ID)
}
\arguments{
\item{SS.table}{result from EMSanova}

\item{del.ID}{id's to combine sum of squares. Use rownames of SS.table}
}
\description{
Pooling nonsignificant interactions to Residuals
}
\examples{
data(film)
anova.result<-EMSanova(thickness~Gate*Operator*Day,data=film,
                        type=c("F","R","R"))
anova.result 
del.ID<-c("Gate:Day","Residuals")
PooledANOVA(anova.result,del.ID)
}
